requirements.txt
asgiref==3.2.7
autopep8==1.5.1
backcall==0.1.0
decorator==4.4.2
Django==3.0.5
django-crispy-forms==1.9.0
entrypoints==0.3
flake8==3.7.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
